Set overall performance based on effort changes
