+++ app/models/event.rb
@@ -39,7 +39,6 @@ class Event < ApplicationRecord
   before_validation :conform_changed_course
   before_save :add_all_course_splits
   after_save :validate_event_group
-  after_touch :set_ranking_data
 
   scope :name_search, -> (search_param) { where("events.name ILIKE ?", "%#{search_param}%") }
   scope :select_with_params, lambda { |search_param|
@@ -172,10 +171,6 @@ class Event < ApplicationRecord
 
   def add_all_course_splits
     splits << course.splits if splits.empty?
-  end
-
-  def set_ranking_data
-    ::Results::SetEventRankingData.perform!(id)
   end
 
   def validate_event_group
