+++ lib/tasks/temp/set_ranking_data.rake
@@ -5,12 +5,12 @@ require "active_record"
 require "active_record/errors"
 
 namespace :temp do
-  desc "sets overall_rank, gender_rank, and status fields for all efforts"
+  desc "sets overall performance and status fields for all efforts"
   task set_ranking_data: :environment do
     Rails.application.eager_load!
 
     Event.find_each do |event|
-      result = ::Results::SetEventRankingData.perform!(event.id)
+      result = ::Results::SetEffortPerformanceData.perform!(event.efforts.pluck(:id))
       puts result.cmd_status
     end
